レコード選択がない状態で変更、削除した際にエラーが出ないよう修正 UndefinedIntを返して受ける形なので、全体の作りは見直す予定
git-svn-id: file:///Users/hiro/repository/beauties@54 b1584ab8-daee-4b4e-b8c6-5a3b018d5d45

+++ Beauties/src/model/action/DeleteRecord.java
@@ -1,5 +1,6 @@
 package model.action;
 
+import model.SystemData;
 import model.db.DbUtil;
 
 import org.eclipse.jface.action.Action;
@@ -19,6 +20,8 @@ public class DeleteRecord extends Action {
 
 	@Override
 	public void run() {
+		if(mCompositeEntry.getSelectedActId() == SystemData.getUndefinedInt())
+			return;
 		if (MessageDialog.openConfirm(mCompositeEntry.getShell(), "確認",
 				"削除していいですか？")) {
 			DbUtil.deleteRecord(mActId);
